refactoring visibility implementation + more tests.
