+++ src/python/pants/testutil/pytest_util.py
@@ -6,16 +6,26 @@ from __future__ import annotations
 from contextlib import contextmanager
 
 
-def assert_logged(caplog, expect_logged: list[tuple[int, str]] | None = None) -> None:
+def assert_logged(
+    caplog, expect_logged: list[tuple[int, str]] | None = None, exclusively: bool = True
+) -> None:
     if not expect_logged:
         assert not caplog.records
         return
 
-    assert len(caplog.records) == len(
-        expect_logged
-    ), f"Expected {len(expect_logged)} records, but got {len(caplog.records)}."
+    if exclusively:
+        assert len(caplog.records) == len(
+            expect_logged
+        ), f"Expected {len(expect_logged)} records, but got {len(caplog.records)}."
+
     for idx, (lvl, msg) in enumerate(expect_logged):
-        log_record = caplog.records[idx]
+        if not exclusively:
+            log_record = next(
+                (log_record for log_record in caplog.records if msg in log_record.message), None
+            )
+            assert log_record is not None, f"Expected log message {msg!r} was not found."
+        else:
+            log_record = caplog.records[idx]
         assert (
             msg in log_record.message
         ), f"The text {msg!r} was not found in {log_record.message!r}."
